Add display.js module
