+++ modules/display.js
@@ -1,0 +1,23 @@
+const listSection = document.querySelector('.booklist');
+const formSection = document.querySelector('.form');
+const contactSection = document.querySelector('.contact-information');
+
+function displayList() {
+  listSection.classList.remove('noshow');
+  formSection.classList.add('noshow');
+  contactSection.classList.add('noshow');
+}
+
+function displayAddNew() {
+  listSection.classList.add('noshow');
+  formSection.classList.remove('noshow');
+  contactSection.classList.add('noshow');
+}
+
+function displayContact() {
+  listSection.classList.add('noshow');
+  formSection.classList.add('noshow');
+  contactSection.classList.remove('noshow');
+}
+
+export {displayList, displayAddNew, displayContact};
